services/user-service/pyproject.toml
[project]
name = "user-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.134.0",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "sqlalchemy>=2.0.47",
]
